pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "local-agent-kit"
version = "0.1.0"
description = "Build local-first AI agents on consumer hardware. Model-agnostic, channel-agnostic, eval-driven."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Benjamin Easington", email = "bene-art@users.noreply.github.com"},
]
keywords = ["ai", "agent", "local", "ollama", "eval", "telegram", "sovereign"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "aiohttp>=3.9",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
telegram = ["aiohttp>=3.9"]
eval = []
all = ["aiohttp>=3.9"]

[project.scripts]
lak = "local_agent_kit.cli:main"

[project.urls]
Homepage = "https://github.com/bene-art/local-agent-kit"
"Reference Agent" = "https://github.com/bene-art/patrick-agent"

[tool.setuptools.packages.find]
where = ["src"]
